+++ src/test/java/com/structurizr/dsl/RelationshipStyleParserTests.java
@@ -101,7 +101,7 @@ class RelationshipStyleParserTests extends AbstractTests {
             parser.parseColour(relationshipStyleDslContext(), tokens("colour", "hex", "extra"));
             fail();
         } catch (Exception e) {
-            assertEquals("Too many tokens, expected: colour <#rrggbb>", e.getMessage());
+            assertEquals("Too many tokens, expected: colour <#rrggbb|color name>", e.getMessage());
         }
     }
 
@@ -111,14 +111,20 @@ class RelationshipStyleParserTests extends AbstractTests {
             parser.parseColour(relationshipStyleDslContext(), tokens("colour"));
             fail();
         } catch (Exception e) {
-            assertEquals("Expected: colour <#rrggbb>", e.getMessage());
+            assertEquals("Expected: colour <#rrggbb|color name>", e.getMessage());
         }
     }
 
     @Test
-    void test_parseColour_SetsTheColour() {
+    void test_parseColour_SetsTheColourWhenUsingAHexColourCode() {
         parser.parseColour(relationshipStyleDslContext(), tokens("colour", "#ff0000"));
         assertEquals("#ff0000", relationshipStyle.getColor());
+    }
+
+    @Test
+    void test_parseColour_SetsTheColourWhenUsingAColourName() {
+        parser.parseColour(relationshipStyleDslContext(), tokens("colour", "yellow"));
+        assertEquals("#ffff00", relationshipStyle.getColor());
     }
 
     @Test